discovery_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
markdown==3.5.2
jinja2==3.1.3
